requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-cloud-compute==1.13.0
google-cloud-storage==2.10.0
google-generativeai==0.3.0
pymongo==4.6.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
aiohttp==3.9.1
httpx==0.25.1
